Refactor Patreon services for better separation of concerns
- Split PatreonService into focused services:
  - PatreonApiClient: Handles HTTP API communication
  - PatreonTokenManager: Manages OAuth token lifecycle
  - PatreonService: Simplified business logic layer
- Add EnableSync configuration flag for background job control
- Remove required validation from optional PatreonOptions tokens
- Update redirect URI for local development environment
