+++ backend/Veloci.Logic/Services/IPatreonService.cs
@@ -4,9 +4,6 @@ namespace Veloci.Logic.Services;
 
 public interface IPatreonService
 {
-    Task<List<PatreonSupporter>> GetCampaignMembersAsync();
-    Task<string?> RefreshAccessTokenAsync(string refreshToken);
-    Task<PatreonTokens?> GetCurrentTokensAsync();
-    Task<string?> GetValidAccessTokenAsync();
-    Task UpdateStoredTokensAsync(string accessToken, string refreshToken, int expiresIn, string? scope = null);
+    Task<List<PatreonCampaign>> GetCampaignsAsync();
+    Task<List<PatreonSupporter>> GetCampaignMembersAsync(string campaignId);
 }
