+++ backend/Veloci.Web/appsettings.json
@@ -16,8 +16,9 @@
   "Patreon": {
     "ClientId": "",
     "ClientSecret": "",
-    "RedirectUri": "https://localhost:7041/patreon/callback",
+    "RedirectUri": "http://localhost:5056/patreon/callback",
     "AccessToken": "",
-    "RefreshToken": ""
+    "RefreshToken": "",
+    "EnableSync": false
   }
 }
